Version 2.6.1
scat.c: Fix bug if -r last command; allow sexigesimal degree search radius
scat.c: Allow rectangular search box (-r radius or -r width height)
newfits.c: Create FITS files as they appear in command line unaffected by
           subsequent command line arguments.  Allow center to be set as
           standard WCSTools coordinate string (ra dec system)

+++ libwcs/ujcread.c
@@ -525,9 +525,9 @@ ujcplate (magetc)
 int magetc;	/* Quality, plate, and magnitude from UJ catalog entry */
 {
     if (magetc < 0)
-	return ((double) (-magetc % 10000000) / 10000);
+	return ( (-magetc % 10000000) / 10000);
     else
-	return ((double) (magetc % 10000000) / 10000);
+	return ( (magetc % 10000000) / 10000);
 }
